requirements.txt
Django>=1.8,<1.9
django-extensions==1.6.1
djangorestframework==3.3.3
mysqlclient==1.3.7
python-dateutil==2.5.1
py-trello==0.4.3
